Moved Bottom Navigation menu composable function to Components.kt, also created a floating button for the the search option
