+++ app/src/main/java/com/example/restaurantproyectdam/MainActivity.kt
@@ -35,6 +35,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import com.example.restaurantproyectdam.ui.components.BottomBar
+import com.example.restaurantproyectdam.ui.components.SearchButton
 import com.example.restaurantproyectdam.ui.navigation.BottomBarOpt
 import com.example.restaurantproyectdam.ui.screens.AdminOrdersScreen
 import com.example.restaurantproyectdam.ui.screens.AdminProductsScreen
@@ -61,38 +63,21 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
 
 @Preview(showBackground = true)
 @Composable
-fun GreetingPreview() {
-    RestaurantProyectDAMTheme {
-        Greeting("Android")
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun ComposeMultiScreenApp(){
-
+fun ComposeMultiScreenApp(){ // MAIN CONTENT
     val navController = rememberNavController()
     Scaffold (
         //color = Color.White
-        bottomBar={ BottomBar(navController = navController) }
+        bottomBar={ BottomBar(navController = navController) },
+        floatingActionButton = { SearchButton(onClick = {}) }
     ) { innerPadding->
         Column(
            modifier = Modifier.padding(innerPadding)
         ){
             SetupNavGraph(navController = navController)
         }
-
-
     }
 }
 
@@ -109,62 +94,4 @@ fun SetupNavGraph (navController: NavHostController){
         composable("menu") { MenuScreen(navController) }
 
     }
-}
-
-
-@Composable
-fun BottomBar(navController:NavHostController){
-    val screens = listOf( //Sealed Classes
-        BottomBarOpt.Menu,
-        BottomBarOpt.Orders,
-        BottomBarOpt.Home,
-        BottomBarOpt.Profile
-    )
-    //Whenever its value is changed, we will be notified about that
-    val navBackStackEntry by navController.currentBackStackEntryAsState()
-    val currentDestionation = navBackStackEntry?.destination // Nullable type // current location
-    //Predefined composable function to add items to the nav bar
-    NavigationBar(){
-        screens.forEach{screen ->
-            AddItem(screen = screen, currentDestination = currentDestionation, navController = navController)
-        }
-    }
-}
-
-@Composable
-fun RowScope.AddItem(// Extends from RowScope
-    screen:BottomBarOpt, //Sealed class
-    currentDestination: NavDestination?, //Nullable type // The current location
-    navController: NavHostController // To navigate to other destinations
-){
-    /*BottomNavigationItem(
-        onClick = { /*TODO*/ },
-        icon = { /*TODO*/ },
-        label = {Text(text = screen.title)}
-    )*/
-    NavigationBarItem(
-        onClick = {
-            navController.navigate(screen.route){
-                //Avoid back stack problem
-                // PopUp the new screen instead of creating a new intace every time
-                popUpTo(navController.graph.findStartDestination().id)
-                //popUpTo(navController.navigate(BottomBarScreen.HomeBB.route))
-                launchSingleTop=true
-                /**
-                 * This allows us to avoid multiple copies of of the same destination
-                 * when reselecting the same item
-                 */
-            }
-        },
-        icon = {
-            when (val icon = screen.icon) {
-                is ImageVector -> Icon(imageVector = icon, contentDescription = "icon")
-                is Int -> Icon(painterResource(icon), contentDescription = "icon", modifier = Modifier.size(30.dp))
-            }
-               },
-        label = {Text(text = screen.title)}, //Text
-        selected = currentDestination?.hierarchy?.any{
-            it.route == screen.route // Checks if is selected
-        } == true
-    )
 }
